Add verify command
- factor out verification from unpack
- fix bug in joining of unpacked split files where hashing would pad all
  parts to 32 byte boundary
