+++ package.json
@@ -26,7 +26,7 @@
   "scripts": {
     "build": "rimraf dist && npx tsc && chmod a+x dist/bin.js",
     "test": "npm run test:unit",
-    "test:unit": "NODE_OPTIONS=--experimental-vm-modules jest --detectOpenHandles --verbose --coverage --roots ./src  --testPathIgnorePatterns \".(e2etest).\"",
+    "test:unit": "NODE_OPTIONS=\"--experimental-vm-modules\" jest --detectOpenHandles --verbose --coverage --roots ./src  --testPathIgnorePatterns \".(e2etest).\"",
     "test:system": "rimraf testing/outputs && NODE_OPTIONS=\"--experimental-vm-modules --enable-source-maps\" jest --verbose --roots ./dist --testPathIgnorePatterns \"\\.(test)\\.\"",
     "lint": "eslint ./src",
     "lint:fix": "eslint --fix ./src",
