+++ src/bin.e2etest.ts
@@ -43,7 +43,10 @@ describe("CLI", function () {
 
   test("verify succeeds", () => {
     expect(() =>
-      execaSync(binPath, ["verify", "path/to/car1", "path/to/car2"])
+      execaSync(binPath, [
+        "verify",
+        "testing/inputs/packs/basic/piece-bafybeie7poazgv7xo4ec7zg6twnuec6qinb3jwqmv6tgu3gbb6olbhlh6q.car",
+      ])
     ).not.toThrow();
   });
 });
