+++ src/cmd/ls.test.ts
@@ -9,4 +9,13 @@ describe("testing ls function", () => {
       )
     ).resolves.not.toThrow();
   });
+
+  test("calling ls verbose does not error", async () => {
+    await expect(
+      ls(
+        "testing/inputs/packs/basic/piece-bafybeie7poazgv7xo4ec7zg6twnuec6qinb3jwqmv6tgu3gbb6olbhlh6q.car",
+        { verbose: true }
+      )
+    ).resolves.not.toThrow();
+  });
 });
