+++ src/cmd/pack.ts
@@ -140,4 +140,6 @@ export default async function pack(
       })
     ) as WritableStream<Buffer>
   );
+
+  console.log(process.getActiveResourcesInfo(), process._getActiveHandlesInfo);
 }
