+++ src/verify.ts
@@ -1,299 +1,145 @@
-import { join } from "node:path";
+/* eslint-disable n/no-unsupported-features/node-builtins */
+import { createHash } from "node:crypto";
+import { createReadStream, WriteStream } from "node:fs";
+import { join, sep } from "node:path";
+import { Readable, Transform, Writable } from "node:stream";
+import { pipeline } from "node:stream/promises";
 
-import { CID } from "multiformats";
+import { createCommPStream } from "@filoz/synapse-sdk/commp";
+import { CarIndexedReader, RawLocation } from "@ipld/car/indexed-reader";
+import { CarIndexer } from "@ipld/car/indexer";
+import { validateBlock } from "@web3-storage/car-block-validator";
+import { createParseStream } from "big-json";
+import { recursive as exporter } from "ipfs-unixfs-exporter";
+import { CID } from "multiformats/cid";
 
-import {
-  SubManifest,
-  SubManifestContentEntry,
-  SuperManifest,
-  SuperManifestContentEntry,
-} from "./manifest.js";
+import { SubManifest } from "./manifest.js";
+import { VerificationFilePart, Verifier } from "./verifier.js";
 
-export interface VerificationFile {
-  hash: string;
-  cid: string;
-  byteLength: number;
-}
+export default async function verify(
+  files: string[],
+  verifier: Verifier,
+  outStreamFactory?: (path: string) => WriteStream,
+  mkdir?: (path: string) => Promise<void>
+): Promise<VerificationFilePart[]> {
+  const fileParts: VerificationFilePart[] = [];
 
-export interface VerificationSplitFile {
-  hash: string;
-  byteLength: number;
-}
+  for (const file of files) {
+    const { stream: commPTransform, getCommP } = createCommPStream();
 
-export interface VerificationFilePart {
-  name: string;
-  byteLength: number;
-  cid: string;
-  originalFileName: string;
-  originalFileHash: string;
-  originalFileByteLength: number;
-}
-
-export type VerificationDirectory = null;
-
-function manifestKeyCheck(
-  sup: SuperManifest,
-  sub: SubManifest,
-  key: keyof SubManifest & keyof SuperManifest
-) {
-  if (sub[key] !== sup[key]) {
-    throw new Error(
-      `Verification failed during manifest '${key}' check, expected '${JSON.stringify(sup[key])}', got '${JSON.stringify(sub[key])}'`
+    const stream = Readable.toWeb(createReadStream(file)).pipeThrough(
+      commPTransform
     );
-  }
-}
 
-class PieceVerifier {
-  #carFile: string;
-  #files: Map<string, VerificationFile>;
-  #dirs: Map<string, VerificationDirectory>;
-  #superManifest: SuperManifest | undefined;
+    const indexer = await CarIndexer.fromIterable(Readable.fromWeb(stream));
+    const index = new Map<string, RawLocation>();
+    const order = [];
+    for await (const { cid, blockLength, blockOffset } of indexer) {
+      const cidStr = cid.toString();
+      index.set(cidStr, { blockLength, blockOffset });
+      order.push(cidStr);
+    }
+    const roots = await indexer.getRoots();
+    const reader = new CarIndexedReader(
+      indexer.version,
+      file,
+      roots,
+      index,
+      order
+    );
 
-  constructor(carFile: string, superManifest?: SuperManifest) {
-    this.#carFile = carFile;
-    this.#files = new Map<string, VerificationFile>();
-    this.#dirs = new Map<string, VerificationDirectory>();
-    this.#superManifest = superManifest;
-  }
-
-  addFile(path: string, info: VerificationFile) {
-    this.#files.set(path, info);
-  }
-
-  addDirectory(path: string) {
-    this.#dirs.set(path, null);
-  }
-
-  verify(subManifest: SubManifest): VerificationFilePart[] {
-    const fileParts: VerificationFilePart[] = [];
-
-    if (this.#superManifest) {
-      // check this piece is part of the supplied dataset
-      for (const key of [
-        "@spec_version",
-        "@spec",
-        "uuid",
-        "name",
-        "description",
-        "version",
-        "license",
-        "project_url",
-        "open_with",
-      ]) {
-        manifestKeyCheck(
-          this.#superManifest,
-          subManifest,
-          key as keyof SubManifest & keyof SuperManifest
-        );
-      }
-
-      // We don't check the unpacked files against the super manifest contents as they will
-      // all be checked against the sub manifest contents.
-
-      // We could choose to verify that the sub manifest and the super manifest contents
-      // are consistent, but that would be a malformed set of manifests, which is not what
-      // we are verifying here, we are verifying that all the files are as declared by the
-      // manifests.
+    const [rootCID] = roots;
+    if (!rootCID) {
+      throw new Error(`No root CID found in CAR: ${file}`);
     }
 
-    if (!subManifest.contents) {
-      // This dataset was packed --lite, so we can't verify the contents and we don't
-      // support split files
-      return [];
-    }
-
-    // Go through the manifest checking that all the files and directories have been added
-    const checkEntries = (
-      path: string[],
-      entries: SubManifestContentEntry[]
-    ) => {
-      for (const entry of entries) {
-        switch (entry["@type"]) {
-          case "file":
-            const actualFile = this.#files.get(join(...path, entry.name));
-            if (!actualFile) {
-              throw new Error(
-                `File '${join(...path, entry.name)} in sub manifest but not extracted from CAR '${this.#carFile}'.`
-              );
-            }
-            if (actualFile.byteLength !== entry.byte_length) {
-              throw new Error(
-                `File '${entry.name}' has size '${String(actualFile.byteLength)}' but sub manifest size is '${String(entry.byte_length)}'.`
-              );
-            }
-            if (actualFile.hash !== entry.hash) {
-              throw new Error(
-                `File '${entry.name}' has hash '${String(actualFile.hash)}' but sub manifest hash is '${String(entry.hash)}'.`
-              );
-            }
-            if (actualFile.cid !== entry.cid) {
-              throw new Error(
-                `File '${entry.name}' has CID '${String(actualFile.cid)}' but sub manifest hash is '${String(entry.cid)}'.`
-              );
-            }
-            // we remove the verified file from the map so we can check for extra files below
-            this.#files.delete(join(...path, entry.name));
-            break;
-
-          case "file-part":
-            const actualFilePart = this.#files.get(join(...path, entry.name));
-            if (!actualFilePart) {
-              throw new Error(
-                `File part '${entry.name}' in sub manifest but not extracted from CAR.`
-              );
-            }
-            if (actualFilePart.byteLength !== entry.byte_length) {
-              throw new Error(
-                `File part '${entry.name}' has size '${String(actualFilePart.byteLength)}' but sub manifest size is '${String(entry.byte_length)}'.`
-              );
-            }
-            if (actualFilePart.cid !== entry.cid) {
-              throw new Error(
-                `File part '${entry.name}' has CID '${String(actualFilePart.cid)}' but sub manifest hash is '${String(entry.cid)}'.`
-              );
-            }
-            fileParts.push({
-              name: join(...path, entry.name),
-              originalFileName: join(...path, entry.original_file_name),
-              originalFileHash: entry.original_file_hash,
-              originalFileByteLength: entry.original_file_byte_length,
-              byteLength: entry.byte_length,
-              cid: entry.cid,
-            });
-
-            // we remove the verified file from the map so we can check for extra ones below
-            this.#files.delete(join(...path, entry.name));
-            break;
-
-          case "directory":
-            const actualDir = this.#dirs.has(join(...path, entry.name));
-            if (!actualDir) {
-              throw new Error(
-                `Directory '${entry.name} in sub manifest but not extracted from CAR.`
-              );
-            }
-            checkEntries([...path, entry.name], entry.contents);
-            // we remove the verified directory from the map so we can check for extra ones below
-            this.#dirs.delete(join(...path, entry.name));
-            break;
+    const entries = exporter(rootCID, {
+      async get(cid) {
+        const block = await reader.get(cid);
+        if (!block) {
+          throw new Error(`Missing block: ${cid.toString()}`);
         }
-      }
-    };
+        await validateBlock(block);
 
-    checkEntries([], subManifest.contents);
+        return block.bytes;
+      },
+    });
 
-    // Any remaining files in the map are spurious and unexpected
-    if (this.#files.size) {
-      throw new Error(
-        `Unpacked files that are not in the sub manifest: ${JSON.stringify(this.#files)}`
-      );
+    const pieceCid = getCommP();
+    if (!pieceCid) {
+      throw new Error("Failed to get CommP from stream");
     }
 
-    // we don't include the root directory in the manifest so remove that
-    this.#dirs.delete(".");
+    const pieceVerifier = verifier.newPieceVerifier(
+      file,
+      rootCID,
+      CID.parse(pieceCid.toString())
+    );
+    let subManifest: SubManifest | undefined;
 
-    // Any remaining directories in the map are spurious and unexpected
-    if (this.#dirs.size) {
-      throw new Error(
-        `Unpacked directories that are not in the sub manifest: ${JSON.stringify(this.#dirs)}`
-      );
-    }
+    for await (const entry of entries) {
+      const filePath = join(...entry.path.split(sep).slice(1));
 
-    return fileParts;
-  }
-}
-
-export class Verifier {
-  #superManifest: SuperManifest | undefined;
-  #pieceCIDs: CID[];
-
-  constructor(manifest?: SuperManifest) {
-    this.#superManifest = manifest;
-    this.#pieceCIDs = [];
-  }
-
-  newPieceVerifier(file: string, payloadCID: CID, pieceCID: CID) {
-    if (this.#pieceCIDs.includes(pieceCID)) {
-      throw new Error(
-        `Piece CID (CommP) '${pieceCID.toString()}' already processed, only provide unique Pieces`
-      );
-    }
-
-    if (this.#superManifest) {
-      // Check that this CAR is a piece CID (CommP) from the super manifest
-      const found = this.#superManifest.pieces.find(
-        (piece) => piece.piece_cid === pieceCID.toString()
-      );
-      if (!found) {
-        throw new Error(
-          `Piece CID (CommP) '${pieceCID.toString()}' does not match any pieces from super manifest '${JSON.stringify(this.#superManifest.pieces)}'`
-        );
-      }
-      if (found.payload_cid !== payloadCID.toString()) {
-        throw new Error(
-          `Payload CID '${payloadCID.toString()}' does not match the piece from super manifest '${JSON.stringify(found)}'`
-        );
-      }
-    }
-
-    this.#pieceCIDs.push(pieceCID);
-
-    return new PieceVerifier(file, this.#superManifest);
-  }
-
-  verifyPieces(splitFiles: Map<string, VerificationSplitFile>) {
-    if (this.#superManifest) {
-      if (this.#superManifest.n_pieces !== this.#pieceCIDs.length) {
-        throw new Error(
-          `Wrong number of CARs (pieces) '${String(this.#pieceCIDs.length)}' provided, expected '${String(this.#superManifest.n_pieces)}'`
-        );
-      }
-
-      if (!this.#superManifest.contents) {
-        return;
-      }
-
-      // Go through the manifest checking that all the split files
-      const checkEntries = (
-        path: string[],
-        entries: SuperManifestContentEntry[]
-      ) => {
-        for (const entry of entries) {
-          switch (entry["@type"]) {
-            case "split-file":
-              const actualFile = splitFiles.get(join(...path, entry.name));
-              if (!actualFile) {
-                throw new Error(
-                  `Split file '${join(...path, entry.name)} in sub manifest but not extracted from ant CAR.`
-                );
-              }
-              if (actualFile.byteLength !== entry.byte_length) {
-                throw new Error(
-                  `File '${entry.name}' has size '${String(actualFile.byteLength)}' but super manifest size is '${String(entry.byte_length)}'.`
-                );
-              }
-              if (actualFile.hash !== entry.hash) {
-                throw new Error(
-                  `File '${entry.name}' has hash '${String(actualFile.hash)}' but super manifest hash is '${String(entry.hash)}'.`
-                );
-              }
-              splitFiles.delete(join(...path, entry.name));
-              break;
-
-            case "directory":
-              checkEntries([...path, entry.name], entry.contents);
-              break;
-          }
+      if (entry.type === "file" || entry.type === "raw") {
+        if (filePath === "manifest.json") {
+          console.log("Sub manifest found");
+          await pipeline(
+            entry.content(),
+            createParseStream().on("data", (data) => {
+              subManifest = data as SubManifest;
+            })
+          );
+          // We don't save the sub manifest to disk so the data is the same as
+          // when it was packed
+          continue;
         }
-      };
+        const hasher = createHash("sha256");
+        const writeStream = outStreamFactory ? outStreamFactory(filePath) : new Writable({
+          write: (_chunk, _encoding, callback) => {
+            callback();
+          },
+        });
+        await pipeline(
+          entry.content(),
+          new Transform({
+            transform(
+              chunk: Uint8Array,
+              _encoding: BufferEncoding,
+              callback: (error?: Error | null, data?: unknown) => void
+            ): void {
+              // If we use the hasher stream then it will round up the byte length
+              // to the nearest block size for the hashing algorithm, so we need to
+              // use the original byte length from the incoming chunk.
+              hasher.update(chunk);
+              this.push(chunk);
+              callback();
+            },
+          }),
+          writeStream
+        );
 
-      checkEntries([], this.#superManifest.contents);
-
-      if (splitFiles.size) {
+        pieceVerifier.addFile(filePath, {
+          hash: hasher.digest("hex"),
+          byteLength: Number(entry.size),
+          cid: entry.cid.toString(),
+        });
+      } else if (entry.type === "directory") {
+        if (mkdir) {
+          await mkdir(filePath);
+        }
+        pieceVerifier.addDirectory(filePath);
+      } else {
         throw new Error(
-          `Unpacked split files that are not in the super manifest: ${JSON.stringify(splitFiles)}`
+          `Unsupported UnixFS type ${entry.type} for path: ${filePath}`
         );
       }
     }
+    await reader.close();
+    if (!subManifest) {
+      throw new Error(`Sub manifest not found in CAR '${file}'`);
+    }
+    fileParts.push(...pieceVerifier.verify(subManifest));
   }
+
+  return fileParts;
 }
